Fixed some issues with PR #14157, kudos to @Nyeriah @Warpten:
* Moved scripts to other source file instead of having them on the instance cpp
* Removed magic number, use of DoAction instead of SetData to do the yelling
* Added spellscript for Remove Kargath's Executioner
* Saved victim GUIDs in instance script instead of checking nearby area
* Returning position as const-reference (by Naios)
