+++ sql/updates/world/2015_MM_DD_NN_world.sql
@@ -1,0 +1,3 @@
+DELETE FROM `spell_script_names` WHERE `spell_id` = 39291;
+INSERT INTO `spell_script_names` (`spell_id`, `ScriptName`) VALUES
+(39291, 'spell_remove_kargath_executioner');
